% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Scatter plot}
\usage{
scatter(data, x, y)
}
\arguments{
\item{data}{df that is being plotted}

\item{x}{x variable}

\item{y}{y variable}
}
\value{
the plot
}
\description{
Scatter plot
}
\examples{
\dontrun{
scatter(df, "income", "total_spent")
}
}
